+++ packages/react-reshaped/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @dfosco/storyboard-react-reshaped
 
+## 4.2.0-alpha.11
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @dfosco/storyboard-react@4.2.0-alpha.11
+
+## 4.2.0-alpha.10
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @dfosco/storyboard-react@4.2.0-alpha.10
+
+## 4.2.0-alpha.9
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @dfosco/storyboard-react@4.2.0-alpha.9
+
+## 4.2.0-alpha.8
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @dfosco/storyboard-react@4.2.0-alpha.8
+
+## 4.2.0-alpha.7
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @dfosco/storyboard-react@4.2.0-alpha.7
+
+## 4.2.0-alpha.6
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @dfosco/storyboard-react@4.2.0-alpha.6
+
+## 4.2.0-alpha.5
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @dfosco/storyboard-react@4.2.0-alpha.5
+
 ## 4.2.0-beta.4
 
 ### Patch Changes